Try check poll before giving up on kill.

+++ libensemble/launcher.py
@@ -51,6 +51,8 @@ def launch(cmd_template, specs=None, **kwargs):
 def killpg(process):
     "Kill the process (and group if it is group leader)."
     try:
+        if process.poll() is not None:
+            return False
         pid = process.pid
         pgid = os.getpgid(pid) if hasattr(os, 'killpg') else -1
         if pgid == pid:
@@ -65,12 +67,14 @@ def killpg(process):
 def terminatepg(process):
     "Send termination signal to the process (and group if it is group leader)"
     try:
+        if process.poll() is not None:
+            return False
         pid = process.pid
         pgid = os.getpgid(pid) if hasattr(os, 'killpg') else -1
         if pgid == pid:
             # Claim in original controller was that we need a wait --
             # have not seen this anywhere else in my searching...
-            time.sleep(0.1)
+            #time.sleep(0.1)
             os.killpg(pgid, signal.SIGTERM)
         elif hasattr(signal, 'CTRL_BREAK_EVENT'):
             # Supposedly does a group terminate for Windows...
